Implemented continuous rendering.

+++ src/main/java/com/laine/casimir/tetris/swing/view/component/panel/JTetrisGamePanel.java
@@ -38,6 +38,7 @@ class JTetrisGamePanel extends JPanel {
     private final TetrisGame tetrisGame = new TetrisGame();
     private final TetrisController tetrisController;
 
+    private long lastDrop;
 
     public JTetrisGamePanel(JFrame frame) {
         this.frame = frame;
@@ -45,11 +46,13 @@ class JTetrisGamePanel extends JPanel {
         this.tetrisController = new TetrisController(tetrisGame);
         this.tetrisController.setTetrisGameListener(tetrisGameListener);
         timer = new Timer(0, e -> {
-            if (!tetrisGame.isGameOver()) {
+            if (!tetrisGame.isGameOver() && System.currentTimeMillis() - lastDrop >= tetrisGame.getDropInterval()) {
                 tetrisController.drop();
-                timer.setDelay(tetrisGame.getDropInterval());
+                lastDrop = System.currentTimeMillis();
             }
+            render();
         });
+        timer.setDelay(0);
         timer.setRepeats(true);
         init();
     }
@@ -97,53 +100,51 @@ class JTetrisGamePanel extends JPanel {
         }
     }
 
+    private void render() {
+        for (int index = 0; index < tetrisSquares.size(); index++) {
+            tetrisSquares.get(index).setBackground(transparent);
+        }
+        final Playfield playfield = tetrisGame.getPlayfield();
+        final List<Square> landedSquares = playfield.getLandedSquares();
+        for (int index = 0; index < landedSquares.size(); index++) {
+            final Square square = landedSquares.get(index);
+            if (square != null) {
+                final Position position = square.getPosition();
+                final int uiSquareIndex = tetrisGrid.getColCount() * position.getY() + position.getX();
+                if (uiSquareIndex >= 0 && uiSquareIndex < tetrisSquares.size()) {
+                    tetrisSquares.get(uiSquareIndex).setBackground(Color.decode(square.getColorHex()));
+                }
+            }
+        }
+        final FallingTetromino fallingTetromino = playfield.getFallingTetromino();
+        if (fallingTetromino != null) {
+            final AbstractTetromino tetromino = fallingTetromino.getTetromino();
+            final Position[] positions = fallingTetromino.getTetromino().getSquarePositions();
+            final Position offset = fallingTetromino.getPosition();
+            for (Position position : positions) {
+                final int uiSquareIndex = tetrisGrid.getColCount() * (position.getY() + offset.getY()) + (position.getX() + offset.getX());
+                if (uiSquareIndex >= 0 && uiSquareIndex < tetrisSquares.size()) {
+                    tetrisSquares.get(uiSquareIndex).setBackground(Color.decode(tetromino.getColorHex()));
+                }
+            }
+        }
+        tetrisGrid.revalidate();
+        tetrisGrid.repaint();
+        holdBoxFragment.setTetromino(tetrisGame.getHoldBox().getTetromino());
+        nextTetrominoFragment.clear();
+        for (int index = 0; index < settings.getNextQueueCount(); index++) {
+            nextTetrominoFragment.addTetromino(tetrisGame.getTetrominoQueue().getPreview(index));
+        }
+        infoFragment.setLines(tetrisGame.getPlayfield().getClearedRows());
+    }
+
     private final TetrisGameListener tetrisGameListener = event -> {
         switch (event) {
-            case GAME_START:
-                setPausePanelVisible(false);
-                break;
             case PAUSE:
                 setPausePanelVisible(true);
                 break;
             case RESUME:
                 setPausePanelVisible(false);
-                break;
-            case DROP:
-                for (int index = 0; index < tetrisSquares.size(); index++) {
-                    tetrisSquares.get(index).setBackground(transparent);
-                }
-                final Playfield playfield = tetrisGame.getPlayfield();
-                final List<Square> landedSquares = playfield.getLandedSquares();
-                for (int index = 0; index < landedSquares.size(); index++) {
-                    final Square square = landedSquares.get(index);
-                    if (square != null) {
-                        final Position position = square.getPosition();
-                        final int uiSquareIndex = tetrisGrid.getColCount() * position.getY() + position.getX();
-                        if (uiSquareIndex >= 0 && uiSquareIndex < tetrisSquares.size()) {
-                            tetrisSquares.get(uiSquareIndex).setBackground(Color.decode(square.getColorHex()));
-                        }
-                    }
-                }
-                final FallingTetromino fallingTetromino = playfield.getFallingTetromino();
-                if (fallingTetromino != null) {
-                    final AbstractTetromino tetromino = fallingTetromino.getTetromino();
-                    final Position[] positions = fallingTetromino.getTetromino().getSquarePositions();
-                    final Position offset = fallingTetromino.getPosition();
-                    for (Position position : positions) {
-                        final int uiSquareIndex = tetrisGrid.getColCount() * (position.getY() + offset.getY()) + (position.getX() + offset.getX());
-                        if (uiSquareIndex >= 0 && uiSquareIndex < tetrisSquares.size()) {
-                            tetrisSquares.get(uiSquareIndex).setBackground(Color.decode(tetromino.getColorHex()));
-                        }
-                    }
-                }
-                tetrisGrid.revalidate();
-                tetrisGrid.repaint();
-                holdBoxFragment.setTetromino(tetrisGame.getHoldBox().getTetromino());
-                nextTetrominoFragment.clear();
-                for (int index = 0; index < settings.getNextQueueCount(); index++) {
-                    nextTetrominoFragment.addTetromino(tetrisGame.getTetrominoQueue().getPreview(index));
-                }
-                infoFragment.setLines(tetrisGame.getPlayfield().getClearedRows());
                 break;
         }
     };
